Force service control block to trigger.
Say an existing service resource "service 'test' do action :enable end" and the desire is
to get the service to enable and start via config-driven-helper::services.

If chef runs and `"services": {"test": ["enable","start"]}` is defined as configuration,
and the service is already enabled but not started, chef considers the service resource
as having nothing to do, so the previous iteration of this recipe would simply do nothing.

This was due to the ruby block only being triggered by the service having been updated
previously in the run.